(:wear:jetpack:cell_grid, :wear:jetpack:shared): move Vector2d to :shared

+++ wear/jetpack/feature/cell_grid/src/main/java/nodomain/pacjo/wear/watchface/feature/cell_grid/Grid2d.kt
@@ -1,5 +1,6 @@
 package nodomain.pacjo.wear.watchface.feature.cell_grid
 
+import nodomain.pacjo.wear.watchface.shared.utils.Vector2.Vector2i
 import kotlin.math.abs
 import kotlin.math.min
 import kotlin.math.pow
@@ -7,7 +8,7 @@ import kotlin.math.sqrt
 
 // extension since I don't feel like this should be a part of the class
 fun <T> Grid2d<T>.setBorder(borderCellValue: T) {
-    val gridCenter = Vector2d(this.width / 2, this.height / 2)
+    val gridCenter = Vector2i(this.width / 2, this.height / 2)
     val radius = min(this.width / 2, this.height / 2)       // TODO: for rectangular devices we should consider cornerRadius instead
 
     // basically we set every cell outside of the circle as BORDER
@@ -24,41 +25,41 @@ fun <T> Grid2d<T>.setBorder(borderCellValue: T) {
 class Grid2d<T>(
     val width: Int,
     val height: Int,
-    init: (position: Vector2d) -> T
+    init: (position: Vector2i) -> T
 ) {
     private val grid =
         MutableList(height) { y ->
             MutableList(width) { x ->
-                init(Vector2d(x, y))
+                init(Vector2i(x, y))
             }
         }
 
-    constructor(size: Int, init: (position: Vector2d) -> T) : this(size, size, init)
+    constructor(size: Int, init: (position: Vector2i) -> T) : this(size, size, init)
 
     /**
      * Indexing helper to skip manual decomposition
      */
-    operator fun get(position: Vector2d): T {
+    operator fun get(position: Vector2i): T {
         return grid[position.y][position.x]
     }
 
     /**
      * Indexing helper to skip manual decomposition
      */
-    operator fun set(position: Vector2d, value: T) {
+    operator fun set(position: Vector2i, value: T) {
         grid[position.y][position.x] = value
     }
 
-    fun isInBounds(position: Vector2d): Boolean {
+    fun isInBounds(position: Vector2i): Boolean {
         return (position.y in 0..<height) && (position.x in 0..<width)
     }
 
     fun forEachPosition(
-        action: (position: Vector2d) -> Unit
+        action: (position: Vector2i) -> Unit
     ) {
         grid.forEachIndexed { y, row ->
             row.forEachIndexed { x, cell ->
-                action(Vector2d(x, y))
+                action(Vector2i(x, y))
             }
         }
     }
